+++ packages/band/tests/distributions/Distributions.test.tsx
@@ -50,7 +50,7 @@ describe('Distributions', () => {
         // the chart should have only bars for 'alpha' - two bars for keys x, y
         expect(screen.getByRole('view-content').querySelectorAll('rect')).toHaveLength(2)
         // the chart should have an axis ready to display 'alpha' and 'beta'
-        const ticks = screen.getAllByRole('tick-label')
+        const ticks = screen.getAllByRole('tick')
         expect(ticks).toHaveLength(2)
         expect(ticks[0].textContent).toEqual('alpha')
         expect(ticks[1].textContent).toEqual('beta')