+++ packages/band/tests/strips/Strips.test.tsx
@@ -52,7 +52,7 @@ describe('Strips', () => {
         const result = screen.getByRole('view-strip')
         expect(result.querySelectorAll('circle')).toHaveLength(22)
         // the chart should have an axis ready to display 'alpha' and 'beta'
-        const ticks = screen.getAllByRole('tick-label')
+        const ticks = screen.getAllByRole('tick')
         expect(ticks).toHaveLength(2)
         expect(ticks[0].textContent).toEqual('alpha')
         expect(ticks[1].textContent).toEqual('beta')