+++ packages/core/src/legends/LegendColorScale.tsx
@@ -74,6 +74,8 @@ const UnthemedLegendColorScale = ({
                     tickSize,
                     labelOffset,
                     labelRotate,
+                    className,
+                    setRole,
                 })}
             </g>
         </g>