+++ packages/core/tests/axes/AxisTicks.test.tsx
@@ -3,6 +3,43 @@ import { Axis, AxisTicks, Chart, View } from '../../src'
 import { chartProps, viewProps } from '../props'
 
 describe('AxisTicks', () => {
+    it('creates ticks with role', () => {
+        render(
+            <Chart {...chartProps}>
+                <View {...viewProps}>
+                    <Axis variant="top">
+                        <AxisTicks variant="top" />
+                    </Axis>
+                </View>
+            </Chart>
+        )
+        const result = screen.getByRole('axis-top').querySelectorAll('g')
+        // the first g inside axis-top should be a group with all ticks
+        expect(result[0].getAttribute('role')).toBe('axis-ticks')
+        // the next g should be a group with one tick
+        expect(result[1].getAttribute('role')).toBe('tick')
+        // elements within the tick group do not need a role
+        expect(result[1].querySelector('line')?.getAttribute('role')).toBeNull()
+        expect(result[1].querySelector('text')?.getAttribute('role')).toBeNull()
+    })
+
+    it('creates ticks without role', () => {
+        render(
+            <Chart {...chartProps}>
+                <View {...viewProps}>
+                    <Axis variant="top">
+                        <AxisTicks variant="top" setRole={false} />
+                    </Axis>
+                </View>
+            </Chart>
+        )
+        const result = screen.getByRole('axis-top').querySelectorAll('g')
+        expect(result[0].getAttribute('role')).toBeNull()
+        expect(result[1].getAttribute('role')).toBeNull()
+        expect(result[1].querySelector('line')?.getAttribute('role')).toBeNull()
+        expect(result[1].querySelector('text')?.getAttribute('role')).toBeNull()
+    })
+
     it('formats tick labels', () => {
         const percentFormat = (v: unknown) => v + '%'
         render(
@@ -14,7 +51,7 @@ describe('AxisTicks', () => {
                 </View>
             </Chart>
         )
-        const result = screen.getAllByRole('tick-label')
+        const result = screen.getAllByRole('tick')
         expect(result[0].textContent).toContain('%')
     })
 
@@ -42,8 +79,7 @@ describe('AxisTicks', () => {
                 </View>
             </Chart>
         )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result).toHaveLength(0)
+        expect(screen.getByRole('axis-top').querySelectorAll('text')).toHaveLength(0)
     })
 
     it('can request specific ticks', () => {
@@ -56,8 +92,7 @@ describe('AxisTicks', () => {
                 </View>
             </Chart>
         )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result).toHaveLength(3)
+        expect(screen.getByRole('axis-top').querySelectorAll('text')).toHaveLength(3)
     })
 
     it('can rotate tick labels', () => {
@@ -70,7 +105,7 @@ describe('AxisTicks', () => {
                 </View>
             </Chart>
         )
-        const result = screen.queryAllByRole('tick-label')
-        expect(result[1].getAttribute('style')).toContain('rotate(')
+        const ticks = screen.getByRole('axis-top').querySelectorAll('text')
+        expect(ticks[0].closest('g')?.getAttribute('style')).toContain('rotate(')
     })
 })